src/requirements.txt
python-dotenv>=0.10.3
azure-identity>=1.16.1
azure-search-documents==11.6.0b5